+++ csvEditorHtml/ui.ts
@@ -1337,6 +1337,16 @@ function displayData(this: any, csvParseResult: ExtendedCsvParseResult | null, c
 
 	})
 
+	{
+		//this is enough (to set it once [after hot creation], because plugins are only disabled but never destroyed)
+		let autoColumnSizePlugin = hot.getPlugin('autoColumnSize')
+		autoColumnSizePlugin.ignoreCellWidthFunc = (value: string) => {
+			const ignoreCommentCellWiths = initialConfig?.autoColumnWithsIgnoreComments ?? true
+
+			return ignoreCommentCellWiths && isCommentCell(value, csvReadConfig)
+		}
+	}
+
 	//@ts-ignore
 	Handsontable.dom.addEvent(window as any, 'resize', throttle(onResizeGrid, 200))