+++ thirdParty/handsontable/handsontable.d.ts
@@ -629,6 +629,8 @@ declare namespace Handsontable {
       samplesGenerator: SamplesGenerator;
       widths: any[];
 
+      ignoreCellWidthFunc: null | (value: string) => boolean;
+
       calculateAllColumnsWidth(rowRange?: number | object): void;
       calculateColumnsWidth(colRange?: number | object, rowRange?: number | object, force?: boolean): void;
       clearCache(columns?: any[]): void;