+++ csvEditorHtml/browser/CHANGELOG.md
@@ -4,6 +4,12 @@ note that the changelog is also present on the website when you click on the ver
 
 ### [Unreleased]
 
+## 1.3.3
+- (same as vs code extension v0.7.3)
+- added option `csv-edit.autoColumnWithsIgnoreComments` to ignore comment cells for auto column sizing
+- `csv-edit.lastRowOrFirstRowNavigationBehavior` default is now `stop` (was `wrap`)
+- `csv-edit.lastColumnOrFirstColumnNavigationBehavior` default is now `stop` (was `wrap`)
+
 ## 1.3.2
 - (same as vs code extension v0.7.2)
 - fixed issue #97 - fixed rows/columns is ignored if `hasHeader` was changed